@@ -5,46 +5,47 @@
   "main": "index.js",
   "scripts": {
     "dev": "cross-env NODE_ENV=development webpack-dev-server  --progress --config ./webpack/config.js",
-    "build:test": "cross-env NODE_ENV=test webpack --open --progress --config ./webpack/config.js",
-    "build:prod": "cross-env NODE_ENV=production webpack --open --progress --config ./webpack/config.js"
+    "build:test": "cross-env NODE_ENV=test webpack --progress --config ./webpack/config.js",
+    "build:prod": "cross-env NODE_ENV=production webpack  --progress --config ./webpack/config.js"
   },
   "author": "lsq",
   "license": "ISC",
   "dependencies": {
-    "ahooks": "^2.2.0",
-    "axios": "^0.19.2",
-    "echarts": "^4.9.0",
-    "html-webpack-plugin": "^4.3.0",
+    "ahooks": "^3.7.7",
+    "axios": "^1.4.0",
+    "echarts": "^5.4.2",
+    "html-webpack-plugin": "^5.5.1",
     "lodash-es": "^4.17.21",
     "lodash-move": "^1.1.1",
-    "react": "^16.13.1",
-    "react-dom": "^16.13.1",
-    "react-router-dom": "^5.2.0",
-    "react-spring": "^8.0.5",
+    "react": "^18.2.0",
+    "react-dom": "^18.2.0",
+    "react-router-dom": "^6.11.2",
+    "react-spring": "^9.7.1",
     "unstated-next": "^1.1.0"
   },
   "devDependencies": {
-    "@types/react": "^16.9.41",
-    "@types/react-dom": "^16.9.8",
-    "@types/react-router-dom": "^5.1.5",
-    "antd": "^4.24.10",
-    "clean-webpack-plugin": "^3.0.0",
-    "copy-webpack-plugin": "^5.1.2",
-    "cross-env": "^7.0.0",
-    "css-loader": "^3.6.0",
-    "dayjs": "^1.8.28",
-    "file-loader": "^6.0.0",
-    "less": "^3.11.3",
-    "less-loader": "^6.1.2",
-    "mini-css-extract-plugin": "^0.9.0",
-    "prettier": "^2.0.5",
+    "@types/react": "^18.2.6",
+    "@types/react-dom": "^18.2.4",
+    "@types/react-router-dom": "^5.3.3",
+    "@webpack-cli/serve": "^2.0.4",
+    "antd": "^5.5.0",
+    "clean-webpack-plugin": "^4.0.0",
+    "copy-webpack-plugin": "^11.0.0",
+    "cross-env": "^7.0.3",
+    "css-loader": "^6.7.4",
+    "dayjs": "^1.11.7",
+    "file-loader": "^6.2.0",
+    "less": "^4.1.3",
+    "less-loader": "^11.1.0",
+    "mini-css-extract-plugin": "^2.7.6",
+    "prettier": "^2.8.8",
     "react-custom-scrollbars": "^4.2.1",
-    "style-loader": "^1.2.1",
-    "ts-loader": "^7.0.5",
-    "typescript": "^3.9.5",
-    "url-loader": "^4.1.0",
-    "webpack": "^4.43.0",
-    "webpack-cli": "^3.3.12",
-    "webpack-dev-server": "^3.11.0"
+    "style-loader": "^3.3.3",
+    "ts-loader": "^9.4.2",
+    "typescript": "^5.0.4",
+    "url-loader": "^4.1.1",
+    "webpack": "^5.83.1",
+    "webpack-cli": "^5.1.1",
+    "webpack-dev-server": "^4.15.0"
   }
 }
