@@ -4,16 +4,16 @@ import React from 'react'
 import {BellOutlined} from '@ant-design/icons'
 import './index.less'
 import {Badge} from 'antd'
-import {useHistory} from 'react-router-dom'
+import {useNavigate} from 'react-router-dom'
 interface IProps {
   number: number
 }
 
 const PageView: React.FC<IProps> = props => {
-  const history = useHistory()
+  const navigate = useNavigate()
   return (
     <Badge count={props.number} offset={[-10, 0]}>
-      <BellOutlined className="notice-tip" onClick={() => history.replace('/need/sys/notice')} />
+      <BellOutlined className="notice-tip" onClick={() => navigate('/need/sys/notice', {replace: true})} />
     </Badge>
   )
 }
