@@ -1,22 +1,22 @@
 /** @format */
 
 import React from 'react'
-import ReactDOM from 'react-dom'
-import App from './app'
+import {createRoot} from 'react-dom/client'
 import {ConfigProvider} from 'antd'
 import Store from '@/store'
-import zh_CN from 'antd/lib/locale-provider/zh_CN'
-import 'antd/dist/antd.css'
+import zhCN from 'antd/locale/zh_CN'
+import 'antd/dist/reset.css'
+import App from './app'
+// import 'antd/dist/antd.css'
 import '@/assets/css/app.less'
 
 const Root = () => {
   return (
     <Store>
-      <ConfigProvider locale={zh_CN}>
+      <ConfigProvider locale={zhCN}>
         <App />
       </ConfigProvider>
     </Store>
   )
 }
-
-ReactDOM.render(<Root />, document.getElementById('root'))
+createRoot(document.getElementById('root') as any).render(<Root />)
