@@ -1,7 +1,7 @@
 /** @format */
 
-import React, {Suspense, lazy, useEffect} from 'react'
-import {Switch, Route, Redirect, useRouteMatch, useHistory} from 'react-router-dom'
+import React, {Suspense, lazy, useEffect, useState} from 'react'
+import {Routes, Route, Navigate, useLocation} from 'react-router-dom'
 import Loading from '@/components/PageLoading/loading'
 import BaseLayout from '@/layouts/basic-layout'
 // import {menuList} from '@/store/menuList'
@@ -9,8 +9,7 @@ import Global from '@/store/global'
 import TagStore from '@/store/tag-view'
 
 const PageViewNeed: React.FC = () => {
-  const match = useRouteMatch()
-  const history = useHistory()
+  const location = useLocation()
   const {addView} = TagStore.useContainer()
   const {token, menuList, getMenuPath} = Global.useContainer()
 
@@ -24,43 +23,34 @@ const PageViewNeed: React.FC = () => {
     return title
   }
   useEffect(() => {
-    // if (menuList?.length) {
-    const unHistory = history.listen(route => {
-      console.log('route', route)
-      addView({pathname: route.pathname, state: {title: getViewName(route.pathname, menuList)}})
-    })
-    return () => {
-      unHistory()
+    if (menuList?.length) {
+      addView({pathname: location.pathname, state: {title: getViewName(location.pathname, menuList)}})
     }
-    // }
-  }, [menuList])
+  }, [menuList, location.pathname])
 
   return token ? (
     <BaseLayout>
       <Suspense fallback={<Loading />}>
-        <Switch>
-          {/* path.join(__dirname, '../src') */}
+        <Routes>
           {menuList.length &&
             menuList.map((item: any, index) => {
+              let Page = lazy(() => import(`@/views${item.component}`))
               // 菜单 并且 需要权限的
               return item.menu_type === 1 && item.is_frame === 0 ? (
-                <Route
-                  key={index}
-                  path={`${match.path}${item.path}`}
-                  component={lazy(() => import(`@/views${item.component}`))}
-                  // component={lazy(() => import(`~/src${item.component}`))}
-                />
+                <Route key={index} path={item.path} element={<Page />} />
               ) : null
             })}
           {/* <Route path={`${match.path}/nav/home`} component={lazy(() => import('@/views/sys/home'))} />
           <Route path={`${match.path}/sys/my`} component={lazy(() => import('@/views/sys/my'))} />
           <Route path={`${match.path}/sys/notice`} component={lazy(() => import('@/views/sys/notice'))} /> */}
-          <Redirect from={match.path} to={'/sys/404'} />
-        </Switch>
+          {/* <Route path="*" element={<Navigate to="/sys/404" replace={true} />} /> */}
+        </Routes>
       </Suspense>
     </BaseLayout>
   ) : (
-    <Redirect to="/noneed/login" />
+    <Routes>
+      <Route path="/*" element={<Navigate to="/noneed/login" replace={true} />} />
+    </Routes>
   )
 }
 
