@@ -1,15 +1,38 @@
 /** @format */
 
 import React, {Suspense, lazy, useEffect} from 'react'
-import {Switch, Route, useRouteMatch, Redirect, useHistory} from 'react-router-dom'
+import {Routes, Route, Navigate, useNavigate} from 'react-router-dom'
 import Loading from '@/components/PageLoading/loading'
 import Global from '@/store/global'
 // import TagStore from '@/store/tag-view'
+let PageLogin = lazy(() => import('@/views/sys/login'))
+
+interface Router {
+  name?: string
+  path: string
+  children?: Array<Router>
+  component: any
+}
 
 const PageViewNoNeed: React.FC = () => {
-  const match = useRouteMatch()
-  const history = useHistory()
+  const navigate = useNavigate()
   const {menuList} = Global.useContainer()
+
+  let router: Array<Router> = menuList.map(item => {
+    return {
+      path: item.path,
+      component: lazy(() => import(`@/views${item.component}`)),
+    }
+  })
+  router.push({
+    path: '/login',
+    component: lazy(() => import('@/views/sys/login')),
+  })
+  router.push({
+    path: '/noneed',
+    component: lazy(() => import('@/views/sys/login')),
+  })
+
   // const {addView} = TagStore.useContainer()
 
   // const getViewName = (pathname, menuList) => {
@@ -22,33 +45,29 @@ const PageViewNoNeed: React.FC = () => {
   //   return title
   // }
 
-  useEffect(() => {
-    const unHistory = history.listen(route => {
-      console.log('noneed', route)
-      // addView({pathname: route.pathname, state: {title: getViewName(route.pathname, menuList)}})
-    })
-    return () => {
-      unHistory()
-    }
-  }, [])
+  // useEffect(() => {
+  //   // const unHistory = navigate.listen(route => {
+  //   //   console.log('noneed', route)
+  //   //   // addView({pathname: route.pathname, state: {title: getViewName(route.pathname, menuList)}})
+  //   // })
+  //   // return () => {
+  //   //   unHistory()
+  //   // }
+  // }, [])
   return (
     <Suspense fallback={<Loading />}>
-      <Switch>
+      <Routes>
         {/* 下面的都是在最外层打开的 */}
-        {menuList.length &&
-          menuList.map((item: any, index) => {
+        {router.length &&
+          router.map((item: any, index) => {
             // 菜单 并且 不需要权限的
             return item.menu_type === 1 && item.is_frame === 1 ? (
-              <Route
-                key={index}
-                path={`${match.path}${item.path}`}
-                component={lazy(() => import(`@/views${item.component}`))}
-              />
+              <Route key={index} path={item.path} element={<item.component />} />
             ) : null
           })}
-        <Route path={`${match.path}/login`} component={lazy(() => import('@/views/sys/login'))} />
-        <Redirect from={match.path} to={`${match.path}/login`} />
-      </Switch>
+        <Route path={`/login`} element={<PageLogin />} />
+        <Route path="/noneed" element={<Navigate to="/noneed/login" replace={true} />} />
+      </Routes>
     </Suspense>
   )
 }
