@@ -3,7 +3,7 @@
 import React from 'react'
 import './index.less'
 import {Form, Input, Button, message} from 'antd'
-import {useHistory} from 'react-router-dom'
+import {useNavigate} from 'react-router-dom'
 import Global from '@/store/global'
 import * as api from '@/api'
 import {setStore} from '@/utils/store'
@@ -13,7 +13,7 @@ interface IProps {
 
 const PageLogin: React.FC<IProps> = props => {
   const [form] = Form.useForm()
-  const hisotry = useHistory()
+  const navigate = useNavigate()
   const {updateToken} = Global.useContainer()
   // const isload = useBoolean()
   const login = () => {
@@ -32,7 +32,7 @@ const PageLogin: React.FC<IProps> = props => {
       if (res.code === 0 && res.data) {
         setStore('user_name', username)
         updateToken('Bearer ' + res.data.token)
-        hisotry.replace('/need/nav/home')
+        navigate('/need/nav/home', {replace: true})
       } else {
         message.error(res.msg)
       }
